+++ Moco/Examples/C++/exampleMocoTrack/exampleMocoTrack.cpp
@@ -160,12 +160,6 @@ void muscleDrivenStateTracking() {
     // problem beyond the MocoTrack interface.
     MocoStudy moco = track.initialize();
 
-    // By default, the MocoTrack tool uses explicit dynamics for the model 
-    // defect constraints. Here, get a reference to the MocoSolver and set the 
-    // dynamics mode to "implicit" instead.
-    MocoCasADiSolver& solver = moco.updSolver<MocoCasADiSolver>();
-    solver.set_dynamics_mode("implicit");
-
     // Get a reference to the MocoControlGoal that is added to every MocoTrack
     // problem by default.
     MocoProblem& problem = moco.updProblem();