+++ Moco/Moco/osimMoco.h
@@ -30,6 +30,7 @@
 #include "MocoCasADiSolver/MocoCasADiSolver.h"
 #include "MocoConstraint.h"
 #include "MocoControlBoundConstraint.h"
+#include "MocoCost/MocoSumSquaredStateCost.h"
 #include "MocoGoal/MocoControlGoal.h"
 #include "MocoGoal/MocoControlTrackingGoal.h"
 #include "MocoGoal/MocoInitialActivationGoal.h"